Sonar Code Smells Cleanup (Redundant casts, simple class names, Stream.toList()) (#2024)
* Code smell updates

* Fix unit tests

+++ CoreInteractiveGraph/src/au/gov/asd/tac/constellation/graph/interaction/plugins/select/FreeformSelectionPlugin.java
@@ -69,10 +69,10 @@ public final class FreeformSelectionPlugin extends SimpleEditPlugin {
         for (i = 0; i < numVertices; i++) {
             j = (i == numVertices - 1) ? 0 : i + 1;
 
-            final float vertY_i = (float) transformedVertices[i * 2 + 1];
-            final float vertX_i = (float) transformedVertices[i * 2];
-            final float vertY_j = (float) transformedVertices[j * 2 + 1];
-            final float vertX_j = (float) transformedVertices[j * 2];
+            final float vertY_i = transformedVertices[i * 2 + 1];
+            final float vertX_i = transformedVertices[i * 2];
+            final float vertY_j = transformedVertices[j * 2 + 1];
+            final float vertX_j = transformedVertices[j * 2];
 
             final boolean belowLowY = vertY_i > yPoint;
             final boolean belowHighY = vertY_j > yPoint;
